requirements.txt
requests >=1.2.0,<1.2.99
python-firebase
google-endpoints